+++ backend/src/modules/references/references.module.ts
@@ -3,9 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Issuance } from './entities/issuance.entity';
 import { IssuanceService } from './services/issuance.service';
 import { IssuanceController } from './controllers/issuance.controller';
+import { Document } from '../documents/entities/document.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Issuance])],
+  imports: [TypeOrmModule.forFeature([Issuance, Document])],
   controllers: [IssuanceController],
   providers: [IssuanceService],
   exports: [IssuanceService],